+++ src/trx/game/objects/creatures/shiva.c
@@ -162,7 +162,7 @@ static void M_TriggerSmoke(const XYZ_32 pos, const bool uw)
     }
 
     spark->scalar = 3;
-    spark->sprite_idx = Object_Get(O_EXPLOSION_1)->mesh_idx;
+    spark->sprite_idx = Object_Get(O_SPARKS_GFX)->mesh_idx;
 
     if (uw) {
         spark->max_y_vel = 0;