+++ src/trx/game/objects/general/smoke_emitter.c
@@ -73,7 +73,7 @@ static void M_Control(const int16_t item_num)
     }
 
     spark->scalar = 3;
-    spark->sprite_idx = Object_Get(O_EXPLOSION_1)->mesh_idx;
+    spark->sprite_idx = Object_Get(O_SPARKS_GFX)->mesh_idx;
     spark->gravity = -8 - (Random_GetControl() & 7);
     spark->max_y_vel = -4 - (Random_GetControl() & 7);
     spark->dst_size.width = (Random_GetControl() & 0x1F) + 128;